some bug fixed

+++ ExecuteAB.py
@@ -23,7 +23,7 @@ The goal is to have the Generator learn this answer key
 
 #Read Data
 n = 1000
-Q = np.genfromtxt('ab_data.csv',delimiter=',')
+Q = np.genfromtxt('ab_data.csv',dtype='int32',delimiter=',')
 Q = np.reshape(Q,(n,5,1))
 #print(Q)
